fix(config): moved the config folder outside auth dir
