+++ src/auth/auth.controller.spec.ts
@@ -6,8 +6,8 @@ import { PrismaService } from '../DAL/prisma/prisma.service';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { AppConfigService } from './config/config.service';
-import { AppConfigModule } from './config/config.module';
+import { AppConfigService } from '../config/config.service';
+import { AppConfigModule } from '../config/config.module';
 
 describe('AuthController', () => {
   let app: INestApplication;
