+++ src/auth/utils/auth.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { PrismaService } from '../../DAL/prisma/prisma.service';
-import { AppConfigService } from '../config/config.service';
+import { AppConfigService } from '../../config/config.service';
 
 @Injectable({})
 export class JwtStrategy extends PassportStrategy(Strategy) {
